requirements.txt
python-telegram-bot==20.7
matplotlib==3.8.4
ccxt==4.1.94
